Issue 19: Initial attempt to fix re-parenting issues
Try to catch WindowHandle during DestroyWindowHandle and restore it CreateWindowHandle. This might cause issues is some (corner?) cases, as the order of messages during creating/destroying window are changed/missed. Also some window properties might not be updated, as VCL requires to recreate window for it.
DoStoreDocState/DoRestoreDocState/OnStoreDocState/OnRestoreDocState are now deprecated (never used).
Tested on D7 and D2009